fix: make riot_id an integer instead of small int since some items have id bigger than 32767

+++ backend/database/models/basic/item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, SmallInteger
+from sqlalchemy import String, Integer, SmallInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database.database import Base
@@ -12,7 +12,7 @@ class Item(Base):
     __tablename__ = "items"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    riot_id: Mapped[int] = mapped_column(SmallInteger, nullable=False)
+    riot_id: Mapped[int] = mapped_column(Integer, nullable=False)
     name: Mapped[str] = mapped_column(String(20), nullable=False)
     description: Mapped[str] = mapped_column(String, nullable=False)
     cost: Mapped[int] = mapped_column(SmallInteger, nullable=False)
